+++ src/LibNoise.NET35/IModule3D.cs
@@ -33,7 +33,7 @@ namespace LibNoise
         /// <param name="y">The input coordinate on the y-axis.</param>
         /// <param name="z">The input coordinate on the z-axis.</param>
         /// <returns>The resulting output value.</returns>
-        float GetValue(float x, float y, float z);
+        double GetValue(double x, double y, double z);
 
         #endregion
     }
